add: JWTによる認証認可処理の追加

+++ src/items/items.service.ts
@@ -1,8 +1,9 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { Item } from '../entities/item.entity';
 import { ItemStatus } from "./item-status.enum";
 import { CreateItemDto } from "./dto/create-item.dto";
 import { ItemRepository } from './item.repository';
+import { User } from 'src/entities/user.entity';
 
 @Injectable()
 export class ItemsService {
@@ -23,12 +24,15 @@ export class ItemsService {
   }
 
 
-  async create(createItemDto: CreateItemDto): Promise<Item> {
-    return await this.itemRepository.createItem(createItemDto);
+  async create(createItemDto: CreateItemDto, user: User): Promise<Item> {
+    return await this.itemRepository.createItem(createItemDto, user);
   }
 
-  async updateStatus(id: string): Promise<Item> {
+  async updateStatus(id: string, user: User): Promise<Item> {
     const item = await this.findById(id);
+    if(item.userId === user.id) {
+      throw new BadRequestException("自身の商品は購入できません。")
+    }
     item.status = ItemStatus.SOLD_OUT
     item.updatedAt = new Date().toISOString();
     this.itemRepository.save(item);
@@ -36,7 +40,11 @@ export class ItemsService {
     return item;
   }
 
-  async delete(id: string): Promise<void> {
+  async delete(id: string, user: User): Promise<void> {
+    const item = await this.itemRepository.findOne({ id });
+    if(item.userId !== user.id) {
+      throw new BadRequestException("他人の商品を削除することはできません");
+    }
     await this.itemRepository.delete({ id });
   }
 }
